+++ MovieBot/MovieBot/States/SearchMovieState.cs
@@ -1,16 +1,11 @@
 ﻿using Microsoft.Bot.Connector;
-using MovieBot.Contract.SearchMovies;
 using MovieBot.Contract;
 using MovieBot.Utility;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
-using System.IO;
 using System.Linq;
 using System.Net;
-using System.Text;
-using System.Web;
-using System.Web.Script.Serialization;
 
 namespace MovieBot.States
 {
